requirements-render.txt
flask==3.0.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
